requirements.txt
ccxt
numpy
pandas
pyyaml
torch
tqdm
requests
